bubblewrap: use new[gu]idmap to setup the user namespace
An unprivileged user can setup only a single mapping in a new user
namespace that maps back to the uid of the user.

The user can be allowed to use more subuids/subgids as configured in the
/etc/subuid and /etc/subgid files.

new[gu]idmap are setuid to run as root and initialize the uid/gid
mappings in the new namespace.  Since they run as root, they can
configure multiple users/groups in the new namespace.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>